+++ src/GLSLVolumeRenderer.cpp
@@ -1407,7 +1407,7 @@ void GLSLVolumeRenderer::drawClippedPolygon( vislib::math::Cuboid<float> boundin
 
     // check for each clip plane
     float vcpd;
-    for( int i = 0; i < this->volClipPlane.Count(); ++i ) {
+	for (int i = 0; i < (int)this->volClipPlane.Count(); ++i) {
         slices.setupSingleSlice( this->volClipPlane[i].PeekComponents(), position.PeekComponents());
         float d = 0.0f;
         vcpd = static_cast<float>(this->volClipPlane[i].PeekComponents()[3]);
